% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightLite.R
\docType{package}
\name{iNZightLite}
\alias{iNZightLite}
\alias{iNZightLite-package}
\title{iNZightLite: a package for running iNZight Lite locally through Docker}
\description{
iNZightLite: a package for running iNZight Lite locally through Docker
}
\author{
Tom Elliott
}
