% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{start}
\alias{start}
\title{Start the Lite docker server}
\usage{
start(port = "3838", name = "inzightlite", open = TRUE)
}
\arguments{
\item{port}{the port to use (the default is 3838, the default for Shiny apps). Only change this if you experience problems.}

\item{name}{the name of the image created. If you change this, be sure to change it in the call to \code{iNZightLite::stop()}, too.}

\item{open}{logical, if set to \code{TRUE}, the app will open in a browser automatically}
}
\value{
the URL to access the Lite app
}
\description{
Start the Lite docker server
}
\examples{
\dontrun{
lite <- iNZightLite::start()
# open in a browser
browseURL(url)
}
}
\author{
Tom Elliott
}
