% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{stop}
\alias{stop}
\title{Stop running Lite process}
\usage{
stop(name = "inzightlite")
}
\arguments{
\item{name}{the name of the container to stop (should be the same used in the call to \code{start()})}
}
\description{
This command will also remove the container.
}
\author{
Tom Elliott
}
